Changing direction if a second button is pressed
If the walk button is pressed in one direction and the opposite direction is pressed, the second button pressed takes priority, and the travel direction resumes if that button is released. This was achieved by keeping track of the initial moving direction as an enum and then overriding it based on the second input.

There's a new struct called InputButton, which we should use for the rest of the inputs in game. It holds whether a button is pressed this frame, held down, or has been released this frame. This is used to keep track of the LEFT and RIGHT buttons as mapped by the InputManager.

I have left lots of comments explaining the code and how it works, and laid out the algorithm step-by-step for you. We will go through it together later!