Made dash faster
Tweaked dash to go faster.
Speed = Distance / Time
dashSpeed = d / startDashTime
so to make it faster, I tweaked dashSpeed from 39 to 60. However because of the equation above, this increases the distance travelled, so I changed startDashTime to 0.3 instead of 0.45 so they were travelling at this speed for a shorter period of time, keeping the distance short.